Change GAME_VERSION to a u32
Doesn't affect anything functionally, but removes the explicit
extra padding we did manually in the encryption key.

+++ src/lib.rs
@@ -43,7 +43,7 @@ pub mod ipc;
 pub mod inventory;
 
 /// Used in the encryption key.
-const GAME_VERSION: u16 = 7201;
+const GAME_VERSION: u32 = 7201;
 
 pub const RECEIVE_BUFFER_SIZE: usize = 32000;
